// packages/shared/src/types/ProjectState.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BranchData } from "./BranchData";

/**
 * Complete project state including all branches
 */
export type ProjectState = { project_name: string, project_path: string, current_branch: string, branches: { [key in string]?: BranchData }, };