// packages/shared/src/types/BranchData.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { VersionInfo } from "./VersionInfo";
import type { WorkingFileInfo } from "./WorkingFileInfo";

/**
 * Information about a branch
 */
export type BranchData = { name: string, description: string | null, versions: Array<VersionInfo>, latest_version: string, parent_branch: string | null, parent_version: string | null, created: string, working_file: WorkingFileInfo | null, };