// packages/shared/src/types/VersionInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Information about a version/commit
 */
export type VersionInfo = { id: string, message: string, author: string | null, timestamp: string, e2k_path: string | null, analyzed: boolean, };