// packages/shared/src/types/ValidationData.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * ETABS file validation data
 */
export type ValidationData = { etabs_installed: boolean, etabs_version: string | null, file_valid: boolean | null, file_path: string | null, file_exists: boolean | null, file_extension: string | null, is_analyzed: boolean | null, validation_messages: Array<string>, };