// packages/shared/src/types/GenerateE2KData.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * E2K generation result data
 */
export type GenerateE2KData = { input_file: string, output_file: string | null, file_exists: boolean, file_extension: string | null, output_exists: boolean | null, generation_successful: boolean | null, file_size_bytes: bigint | null, generation_time_ms: bigint | null, messages: Array<string>, };